API latest release (#31)
* Gitignore

* Initial empty API

* Endpoints

Created API endpoint structure

* Schemas

Uploaded empty schemas for every API endpoint request

* Tools

Implemented some tools that will be useful. JSON schema still to complete

* JSON Schema

Updates JSON Schemas

* Authors, Encryption

Added author to files and implemented encryption

* Added some functions

* Endpoints complete

Initial version of the API with all endpoints implemented

* Update JSON Schema

Updated JSON schema to allow for title and description to have punctuation

* Updated to Server

Updated script to be run on server with WSGI

* Fixes

Resolves #17
This commit fixes an issue with the login API that was causing it to reject special characters passed by the frontend. The issue was caused by a validation check that was too strict.

* Bug fixes

Fixed bugs with logging in users

* Bug fixes

The API will now return the itemID when creating a new Item. Bug fixes with creating an auction

* Updated getBid

Updated getBid to allow to get the highest bid for an auction

* Update veriyIban to POST

Updated the verifyIban endpoint to use POST instead of GET

* Bug fixes

* Merge branch 'API_Main' of https://github.com/TheRealTimo/CarAuctionApp into API_Main

* PR

Co-authored-by: Timo <31155789+TheRealTimo@users.noreply.github.com>